refactor idle
