+++ web-ui/src/features/board/view-model/variants/idle/use-go-to-add-sticker.ts
@@ -1,0 +1,17 @@
+import type React from "react";
+import type { ViewModelParams } from "../../view-model-params";
+import { goToSticker } from "../add-sticker";
+
+export function useGoToAddSticker({ setViewState }: ViewModelParams) {
+  const handleKeyDown = (e: React.KeyboardEvent) => {
+    if (e.code === "KeyS") {
+      setViewState(goToSticker());
+    }
+  };
+
+  const moveToStickerState = () => {
+    setViewState(goToSticker());
+  };
+
+  return { handleKeyDown, moveToStickerState };
+}
